* uninitialized read checking * mf-runtime API tweak
[gcc/ChangeLog.tree-ssa]
2003-01-29  Frank Ch. Eigler  <fche@redhat.com>

        * tree-mudflap.c (mx_xfn_indirect_ref): Use a stack of flags for
        tracking pointer deference reads vs writes.
        (mf_build_check_statement_for): Pass access-type value to __mf_check.
        * c-mudflap.c (mflang_register_call): Adapt to mf-runtime.h API change.

[gcc/cp/ChangeLog.tree-ssa]
2003-01-29  Frank Ch. Eigler  <fche@redhat.com>

        * cp-mudflap.c (mflang_register_call): Adapt to mf-runtime.h API
	change.

[libmudflap/ChangeLog]
2003-01-29  Frank Ch. Eigler  <fche@redhat.com>

        * Makefile.am (TESTS_ENVIRONMENT): Remove redundant "-mode-check".
        (TESTS): Add fail22 and pass29 tests.
        * mf-runtime.h.in: Change API to take void*/size_t region parameters.
        Add new access-type parameter for __mf_check.  Move __MF_VIOL* out.
        * mf-impl.h: Corresponding changes.  Update CLAMP* macros for void*
        values.  Move __MF_VIOL* here.
        * mf-runtime.c (*): Adapt to void*/size_t API in mf-runtime.h.
        (check_initialization): New field in __mf_opts.  Default off.
        (read_count,write_count): New fields in __mf_object.
        (__mf_check): Implement basic initialization checking.
        (__mf_insert_new_object): Assume STATIC|GUESS regions are initialized.
        (__mf_describe_object): Print new fields.
        (__mf_violation): Identify check/read vs. check/write in messages.
        * test/pass29-frag.c, test/fail22-frag.c: Basic tests for new
        "-check-initialized" mudflap option.
        * test/pass25-frag.c, test/fail21-frag.c: Adapt to API changes.
        * mf-hooks.c (MF_VALIDATE_EXTENT): Add new access-type parameter.
        Drop __FILE__/__LINE__ hack.  Update callers.
        (*): Adapt to new mf-runtime.h API.
        * Makefile.in: regenerated.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/tree-ssa-20020619-branch@62101 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libmudflap/test/fail21-frag.c
@@ -2,6 +2,6 @@ int foo = 5;
 int *bar = & foo;
 /* Make an access here to get &foo into the lookup cache.  */
 *bar = 5;
-__mf_watch ((uintptr_t) & foo, sizeof(foo));
+__mf_watch (& foo, sizeof(foo));
 /* This access should trigger the watch violation.  */
 *bar = 10;
